Refactored structure to have a smaller, more basic, more generic core.
Removed the implicit cell.  Syntax can be misleading, lead to mistakes, and is extra work to upkeep.  (Perhaps it can be re-implemented later as a side library.)

`Cell` is now an abstract class so that the "smart" part that calculates the value (the "backend") is implementable in a more general and simpler way.
Two such implementations are the `SettableCell` (cell managed/set by non-frp code) and the cell function (`cf`) where the backend is hidden and only the face of the cell is returned.